Refactor, fixes and maximized option for title tab (#40)
* move definitions of switcher structs to separate files

* remove useless comment

* cleanup and silence warnings

* add generic switch

* removed delay on main thread startup as it served no purpose

* clean up ultility.hpp

* do not save switchStr to file as it could cause issues when scenes are renamed

* rename sceneRoundTrip to sceneSequence

* add option to switch if window is maximized to title tab

+++ src/switch-priority.cpp
@@ -1,10 +1,11 @@
 #include <algorithm>
+
 #include "headers/advanced-scene-switcher.hpp"
 
 void SceneSwitcher::on_threadPriority_currentTextChanged(const QString &text)
 {
-	if (loading ||
-	    ui->threadPriority->count() != switcher->threadPriorities.size())
+	if (loading || ui->threadPriority->count() !=
+			       (int)switcher->threadPriorities.size())
 		return;
 
 	std::lock_guard<std::mutex> lock(switcher->m);
